Fixed bugs in umap_spike_scatter

+++ NeuronCluster/umap_spike_scatter.py
@@ -22,39 +22,45 @@ cluster_num = int(params[0])
 
 
 def umap_plots(data_dir, electrode_num):
-    pca_waveforms = np.load(data_dir + \
-            '/spike_waveforms/electrode{}/pca_waveforms.npy'.format(electrode_num))
+    # If processing has happened, the file will exist
+    pca_file = data_dir + \
+                '/spike_waveforms/electrode{}/pca_waveforms.npy'.format(electrode_num)
 
-    umap_waveforms = umap.UMAP(n_components = 2).\
-            fit_transform(pca_waveforms[:,:20])
-    
-    clustering_results = [np.load(data_dir + \
-            '/clustering_results/electrode{0}/clusters{1}/predictions.npy'.\
-            format(electrode_num, cluster)) for cluster in \
-            range(2,cluster_num+1)] 
-    
-    print('Processing for Electrode {} complete'.format(electrode_num))
+    if os.path.isfile(pca_file):
 
-    for cluster in range(2,cluster_num+1):
-        fig1, ax1 = plt.subplots()
-        scatter = ax1.scatter(umap_waveforms[:,0],umap_waveforms[:,1],\
-                c = clustering_results[cluster-2], s = 2, cmap = 'jet')
-        legend = ax1.legend(*scatter.legend_elements())
-        ax1.add_artist(legend)
-        fig1.savefig(data_dir + \
-            '/Plots/{0}/Plots/{1}_clusters_waveforms_ISIs/cluster{1}_umap.png'.\
-            format(electrode_num, cluster), 
-            dpi = 300)
-        plt.close(fig1)
+        pca_waveforms = np.load(data_dir + \
+                '/spike_waveforms/electrode{}/pca_waveforms.npy'.format(electrode_num))
 
-        nbins = np.min([100,int(umap_waveforms.shape[0]/100)])
-        fig2, ax2 = plt.subplots()
-        ax2.hexbin(umap_waveforms[:,0],umap_waveforms[:,1], gridsize = nbins)
-        fig2.savefig(data_dir + \
-            '/Plots/{0}/Plots/{1}_clusters_waveforms_ISIs/cluster{1}_umap_hist.png'.\
-            format(electrode_num, cluster),
-            dpi = 300)
-        plt.close(fig2)
+        umap_waveforms = umap.UMAP(n_components = 2).\
+                fit_transform(pca_waveforms[:,:20])
+        
+        clustering_results = [np.load(data_dir + \
+                '/clustering_results/electrode{0}/clusters{1}/predictions.npy'.\
+                format(electrode_num, cluster)) for cluster in \
+                range(2,cluster_num+1)] 
+        
+        print('Processing for Electrode {} complete'.format(electrode_num))
+
+        for cluster in range(2,cluster_num+1):
+            fig1, ax1 = plt.subplots()
+            scatter = ax1.scatter(umap_waveforms[:,0],umap_waveforms[:,1],\
+                    c = clustering_results[cluster-2], s = 2, cmap = 'jet')
+            legend = ax1.legend(*scatter.legend_elements())
+            ax1.add_artist(legend)
+            fig1.savefig(data_dir + \
+                '/Plots/{0}/Plots/{1}_clusters_waveforms_ISIs/cluster{1}_umap.png'.\
+                format(electrode_num, cluster), 
+                dpi = 300)
+            plt.close(fig1)
+
+            nbins = np.min([100,int(umap_waveforms.shape[0]/100)])
+            fig2, ax2 = plt.subplots()
+            ax2.hexbin(umap_waveforms[:,0],umap_waveforms[:,1], gridsize = nbins)
+            fig2.savefig(data_dir + \
+                '/Plots/{0}/Plots/{1}_clusters_waveforms_ISIs/cluster{1}_umap_hist.png'.\
+                format(electrode_num, cluster),
+                dpi = 300)
+            plt.close(fig2)
 
 for electrode_num in trange(len(os.listdir(data_dir + '/clustering_results'))):
     umap_plots(data_dir, electrode_num)
